backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-allauth==0.57.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-cloudinary-storage==0.3.0
cloudinary==1.37.0
Pillow==10.1.0
python-decouple==3.8
django-filter==23.5
django-extensions==3.2.3
reportlab==4.0.7
requests==2.31.0
python-mpesa==0.4.9
